pip_requirements.txt
absl-py==0.15.0
anyascii==0.3.1
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
Babel==2.9.1
backports.zoneinfo==0.2.1
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
clang==5.0
click==8.1.2
colorama==0.4.4
coqpit==0.0.15
coqui-trainer==0.0.5
cycler==0.11.0
Cython==0.29.28
dataclasses==0.6
dateparser==1.0.0
decorator==5.1.1
deepspeech==0.9.3
deprecation==2.1.0
Distance==0.1.3
docopt==0.6.2
filelock==3.6.0
Flask==2.1.1
flatbuffers==1.12
fonttools==4.32.0
fsspec==2022.3.0
future==0.17.1
g2p-en==2.1.0
g2pM==0.1.2.5
gast==0.4.0
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gruut==2.0.4
gruut-ipa==0.10.1
gruut-lang-cs==2.0.0
gruut-lang-de==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.0
gruut-lang-it==2.0.0
gruut-lang-nl==2.0.0
gruut-lang-pt==2.0.0
gruut-lang-ru==2.0.0
gruut-lang-sv==2.0.0
h5py==3.1.0
halo==0.0.31
huggingface-hub==0.0.8
idna==3.3
importlib-metadata==4.11.3
inflect==5.5.2
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.1
joblib==0.14.1
jsonlines==1.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
librosa==0.8.0
llvmlite==0.36.0
log-symbols==0.0.14
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
mecab-python3==1.0.3
networkx==2.6.3
nltk==3.7
num2words==0.5.10
numba==0.53.0
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
picovoice==2.1.0
Pillow==9.1.0
playsound==1.3.0
pooch==1.6.0
protobuf==3.20.0
pvporcupine==2.1.0
pvrhino==2.1.0
py-picotts==0.1.2
pyaml==19.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.21.0
pycparser==2.21
pynndescent==0.5.6
pyparsing==3.0.8
pypinyin==0.46.0
pysbd==0.3.4
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyworld==0.3.0
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
scikit-learn==0.22.2.post1
scipy==1.7.3
six==1.15.0
sklearn-crfsuite==0.3.6
snips-nlu==0.20.2
snips-nlu-de @ https://github.com/snipsco/snips-nlu-language-resources/releases/download/snips_nlu_de-0.2.3/snips_nlu_de-0.2.3.tar.gz
snips-nlu-metrics==0.14.6
snips-nlu-parsers==0.4.3
snips-nlu-utils==0.9.1
sounddevice==0.4.4
SoundFile==0.10.3.post1
spinners==0.0.24
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow-addons==0.16.1
tensorflow-estimator==2.8.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
TextGrid==1.5
torch==1.11.0
torchaudio==0.11.0
tqdm==4.64.0
typeguard==2.13.3
typing-extensions==3.7.4.3
tzdata==2022.1
tzlocal==4.2
umap-learn==0.5.1
Unidecode==1.3.4
unidic-lite==1.0.8
urllib3==1.26.9
webrtcvad==2.0.10
Werkzeug==2.1.1
Wikipedia-API==0.5.4
wrapt==1.12.1
zipp==3.8.0
